+++ pkg/helper/k8smeta/k8s_meta_deferred_deletion_meta_store_test.go
@@ -32,7 +32,6 @@ func TestDeferredDeletion(t *testing.T) {
 			Raw: pod,
 		},
 	}
-	time.Sleep(10 * time.Millisecond)
 	cache.lock.RLock()
 	if _, ok := cache.Items["default/test"]; !ok {
 		t.Errorf("failed to add object to cache")
